add bit op

+++ mpi_fetch_data/mpi_processor.h
@@ -13,7 +13,7 @@ struct mpiProcessor {
     int serverfd;
     char type; // 0 is reader,1 is writer,2 is producer
     struct tokenInfo token;
-    uint8_t *store;
+    uint32_t *store;
 };
 int mpiProcessorInit(struct mpiProcessor *mp,char *remote_host,int port,int type);
 void mpiProcessRun(struct mpiProcessor *mp);
